requirements.txt
uvicorn~=0.20.0
fastapi~=0.92.0
sqlalchemy~=2.0.4
python-dotenv
psycopg2
alembic~=1.9.4
passlib~=1.7.4
python-jose
python-multipart
pydantic~=1.10.5